Chart Display: Move default implementations into separate package
