+++ src/main/java/net/imagej/defaultplot/DefaultLineSeries.java
@@ -1,4 +1,7 @@
-package net.imagej.plot;
+package net.imagej.defaultplot;
+
+import net.imagej.plot.LineSeries;
+import net.imagej.plot.SeriesStyle;
 
 import java.util.Collection;
